feat: 494 — Chebyshev realignment retires four plan-failure regressions
Position::distance_to now returns Chebyshev (king-move) distance — the
substrate-correct metric matching 8-direction movement cost per
ai/pathfinding.rs::heuristic. Pre-494 it returned Euclidean, producing
a perception/movement mismatch where cats scored targets at a metric
they couldn't actually traverse (diagonal-of-1 reads as 1.41 under
Euclidean, but costs 1 step to walk; Chebyshev reads it as 1).

Fixes (494 plan-failure rows; pre-fix → post-fix per seed 42 soak):

- SearchPrey "no scent found":   1 → 167 → 9
- EngagePrey "no prey target":  14 → 205 → 10
- TravelTo(PatrolZone) stuck:    0 → 530 → 0
- MentorCat Incapacitated:       0 → 449 → 0

Mechanism:

- physical.rs: distance_to → Chebyshev; tile_distance_squared →
  Chebyshev² (preserves Ord for min_by_key); euclidean_distance added
  as escape hatch for sites that genuinely measure radial physics
  (scent diffusion, ward glow falloff).
- prey/cat/fox/carcass scent map internal gates flipped from Manhattan
  to Chebyshev for caller-symmetry under the new default.
- goap.rs::resolve_zone_position: perimeter_offset_position helper
  adds reachability gate to PatrolZone + RestingSpot branches. Pre-494
  these constructed a blind +x offset off the nearest store and landed
  on water/wall when stores hugged map edges. Helper rotates through
  4 cardinal offsets and returns the first passable in-bounds tile.
- Completes the in-progress 492 distance-metric migration: ~70
  manhattan_distance call sites in goap.rs routed to distance_to
  (comparisons against f32 range constants) or tile_distance_squared
  (min_by_key / sort_by_key picks).

Test updates (substrate-effect calibration, not functional regressions):
- components::physical::tests::position_distance: asserts Chebyshev 4
  (was Euclidean 5.0).
- steps::hawk::tests::soar_moves_diagonally_then_advances: target
  bumped (2,2) → (3,3) so two ticks are still needed under Chebyshev.
- ai::pathfinding::tests::avoids_water_diagonal_tries_cardinal: uses
  euclidean_distance for the "made geometric progress" assertion
  (Chebyshev cost metric doesn't strictly decrease on cardinal-toward-
  diagonal steps).
- ai::dses::caretake_target::tests::isolated_kitten_beats_co_located_sibling:
  kitten positions bumped to radius 4 so the ISOLATION_RADIUS=3 gate
  still separates lonely from co-located under Chebyshev.
- scenarios::prey_byproduct_spawn rat tick budget 200 → 800; bird
  test counts cat-only kills via inventory rather than the global
  ByproductSpawned canary.

Net result: colony survived 91k ticks vs baseline 59k (+54%, 4
seasons vs 2). Substrate-effect follow-on regressions
(TravelTo(CarcassPile) pathing 0 → 1099, HoldUntilSafe timeout
0 → 670, welfare composition drift) documented in ticket 494's
post-fix follow-ons section.

+++ src/components/prey.rs
@@ -79,7 +79,7 @@ pub struct PreyConfig {
     pub item_kind: ItemKind,
     pub flee_speed: u32,
     pub graze_cadence: u64,
-    pub alert_radius: i32,
+    pub alert_radius: f32,
     pub freeze_ticks: u64,
     pub catch_difficulty: f32,
     pub flee_strategy: FleeStrategy,
